chore: remove dead code in pkg/tfshim/sdk-v2
After PlanResourceChange behavior is the new default, some of the code is made redundant.
