+++ pkg/tfshim/sdk-v2/upgrade_state.go
@@ -10,7 +10,6 @@ import (
 	"github.com/hashicorp/go-cty/cty/msgpack"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 func upgradeResourceStateGRPC(
@@ -85,90 +84,4 @@ func extractSchemaVersion(meta map[string]any) (int64, bool, error) {
 		return 0, true, err
 	}
 	return v, true, nil
-}
-
-func upgradeResourceState(ctx context.Context, t string, p *schema.Provider, res *schema.Resource,
-	instanceState *terraform.InstanceState,
-) (*terraform.InstanceState, error) {
-	if instanceState == nil {
-		return nil, nil
-	}
-
-	rawState := instanceState.RawState
-
-	// If RawState is not set but attributes is, we need to hydrate RawState
-	// from attributes.
-	if rawState.IsNull() {
-		// We default to assuming that the old state has the same shape as the new
-		// resource.
-		typ := res.CoreConfigSchema().ImpliedType()
-
-		// Find the version, if present, to deserialize instanceState into.
-		version, hasVersion, err := extractSchemaVersion(instanceState.Meta)
-		if err != nil {
-			return nil, err
-		}
-
-		// If we have a version, we use the schema shape that matches the version
-		// specified.
-		if hasVersion {
-			for _, t := range res.StateUpgraders {
-				if t.Version == int(version) {
-					typ = t.Type
-					break
-				}
-			}
-		}
-
-		rawState, err = instanceState.AttrsAsObjectValue(typ)
-		if err != nil {
-			return nil, fmt.Errorf("state from attributes: %w", err)
-		}
-	}
-
-	if state := rawState.AsValueMap(); !has(state, "id") {
-		state["id"] = cty.StringVal(instanceState.ID)
-		rawState = cty.ObjectVal(state)
-	}
-
-	v, newMeta, err := upgradeResourceStateGRPC(ctx, t, res, rawState, instanceState.Meta, p.GRPCProvider())
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert the value back to an InstanceState.
-	newState, err := res.ShimInstanceStateFromValue(v)
-	if err != nil {
-		return nil, err
-	}
-
-	newState.RawConfig = instanceState.RawConfig
-	newState.RawState = v
-	newState.Meta = newMeta
-	newState.ID = instanceState.ID
-
-	// If state upgraders have modified the ID, respect the modification.
-	if updatedID, ok := findID(v); ok {
-		newState.ID = updatedID
-	}
-	return newState, nil
-}
-
-func findID(v cty.Value) (string, bool) {
-	if !v.Type().IsObjectType() {
-		return "", false
-	}
-	id, ok := v.AsValueMap()["id"]
-	if !ok {
-		return "", false
-	}
-	if !id.Type().Equals(cty.String) || id.IsNull() {
-		return "", false
-	}
-	return id.AsString(), true
-}
-
-func has[K comparable, V any](m map[K]V, k K) bool {
-	_, ok := m[k]
-	return ok
 }
